+++ app/gui/src/dashboard/modals/PlanDowngradedModal.tsx
@@ -5,9 +5,8 @@ import { Text } from '#/components/Text'
 import { useEventCallback } from '#/hooks/eventCallbackHooks'
 import { useLocalStorageState } from '#/hooks/localStoreState'
 import LocalStorage from '#/utilities/LocalStorage'
-import { DAY_MS, HOUR_MS } from '#/utilities/time'
+import { DAY_MS, HOUR_MS, HOURS_PER_DAY, MINUTE_MS, useCurrentTimestamp } from '#/utilities/time'
 import { useText } from '$/providers/react'
-import { useEffect, useState } from 'react'
 import { z } from 'zod'
 
 /** Props for a {@link PlanDowngradedModal}. */
@@ -36,15 +35,8 @@ const MIN_SHOW_INTERVAL = HOUR_MS
 /** Maximum amount of time that must pass before the modal is shown again. */
 const MAX_SHOW_INTERVAL = DAY_MS * 5
 
-function useCurrentTimestamp(refreshInterval: number) {
-  const [timestampValue, setTimestampValue] = useState(Date.now())
-  useEffect(() => {
-    const interval = setInterval(() => setTimestampValue(Date.now()), refreshInterval)
-    return () => clearInterval(interval)
-  }, [refreshInterval])
-  return timestampValue
-}
-
+/** Hook for accessing or setting storage state related to downgrade modal. */
+// eslint-disable-next-line react-refresh/only-export-components
 export function useDowngadeModalState() {
   const [storageState, setStorageState] = useLocalStorageState('downgradeModal')
 
@@ -66,11 +58,11 @@ export function PlanDowngradedModal(props: PlanDowngradedModalProps) {
   const { lastShownTimestamp, markAsShown } = useDowngadeModalState()
 
   // Progress time reference every minute, so we can show the modal as time goes on.
-  const referenceNowTime = useCurrentTimestamp(60000)
+  const referenceNowTime = useCurrentTimestamp(MINUTE_MS)
 
   const msToDeadline = deletionDeadlineTimestamp - referenceNowTime
   const daysLeft = Math.floor(msToDeadline / DAY_MS)
-  const hoursLeft = Math.floor(msToDeadline / HOUR_MS) % 24
+  const hoursLeft = Math.floor(msToDeadline / HOUR_MS) % HOURS_PER_DAY
 
   // Show alert again if the time elapsed since last showing is greater than the time left to delete,
   // That way the alerts become more frequent as the deadline approaches. Limited by set min/max range.
